Added laplace encoding distribution.
Fixed bug in normal distribution which caused it to be offset by 0.5 from the mean.

Fixed bug in ImageLayerCompressor::calcDists that caused it to not correctly process all input.

Fixed bug in ImageLayerCompressor where it would not correctly use the scale parameter when calculating pixel offsets (was using fixed 1 pixel offsets instead).